don't double cast to u8

+++ src/lib.rs
@@ -200,7 +200,7 @@ pub fn gen_register_r(r: &Register, d: &Defaults) -> Vec<Tokens> {
         }
 
         let name = Ident::new(field.name.to_snake_case());
-        let offset = u8::from(field.bit_range.offset as u8);
+        let offset = field.bit_range.offset as u8;
 
         let width = field.bit_range.width;
